Add enhanced support to the fluid sampler
"Enhanced features" (the ability to specify multiple observation files, and
observables that are functions of the state rather than direct
measurements of species) were previously added to the ABC sampler. This
update allows the fluid sampler to take advantage of the same features.

There is no new sampler; users should just specify "fluid" or "ODE" as
before. Performance does not seem to have suffered (for a small example,
at least) but it may be worth keeping in mind that the new likelihood
computation includes function calls (for mapping the observables) compared
to just numpy array operations before (although the ODE solution might be
eclipsing that cost anyway).